added authentication and authorization for the endpoints

+++ Backend/Application/Features/Auth/Dtos/TokenDto.cs
@@ -3,7 +3,7 @@ namespace Application.Features.Auth.Dtos
     public class TokenDto
     {
         public string AccessToken { get; set; }
-        public int ExpireInDays { get; set; }
+        public double ExpireInDays { get; set; }
     }
 
     public class NonceDto
